refac: datalist refac
